apps/mush-worker/pyproject.toml
[project]
name = "mush-worker"
version = "0.1.0"
description = ""
authors = [
    { name = "Daniel", email = "126309503+danielxfeng@users.noreply.github.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "timm (>=1.0.19,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "sentry-sdk[fastapi] (>=2.35.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"
types-requests = "^2.32.4.20250809"
types-redis = "^4.6.0.20241004"
pytest-asyncio = "^1.1.0"
pre-commit = "^4.3.0"


[tool.mypy]
python_version = '3.13'
disallow_any_unimported = false
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
disallow_untyped_defs = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
mypy_path = './stubs'

[tool.isort]
profile = 'black'
src_paths = ['mush_worker', 'tests']
line_length = 119

[tool.black]
line-length = 119
target-version = ['py313']
include = '\.pyi?$'
